Renamed ThemeScene2 to ThemeScene
