+++ packages/edge-login-ui-rn/src/components/scenes/WaitScene.tsx
@@ -4,7 +4,7 @@ import { cacheStyles } from 'react-native-patina'
 
 import { Theme, ThemeProps, withTheme } from '../services/ThemeContext'
 import { EdgeText } from '../themed/EdgeText'
-import { ThemedScene } from '../themed/ThemedScene2'
+import { ThemedScene } from '../themed/ThemedScene'
 
 const loader = require('../../assets/safeLoader.gif')
