+++ packages/edge-login-ui-rn/src/components/scenes/existingAccout/SecurityAlertsScene.tsx
@@ -16,7 +16,7 @@ import { Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
 import { IconHeaderRow } from '../../themed/IconHeaderRow'
 import { LinkRow } from '../../themed/LinkRow'
 import { PrimaryButton } from '../../themed/ThemedButtons'
-import { ThemedScene } from '../../themed/ThemedScene2'
+import { ThemedScene } from '../../themed/ThemedScene'
 import { MessageText, Warning } from '../../themed/ThemedText'
 
 interface OwnProps {}
