+++ packages/edge-login-ui-rn/src/components/scenes/newAccount/NewAccountReviewScene.tsx
@@ -15,7 +15,7 @@ import { EdgeText } from '../../themed/EdgeText'
 import { Fade } from '../../themed/Fade'
 import { FormError } from '../../themed/FormError'
 import { SecondaryButton } from '../../themed/ThemedButtons'
-import { ThemedScene } from '../../themed/ThemedScene2'
+import { ThemedScene } from '../../themed/ThemedScene'
 
 interface OwnProps {}
