+++ packages/edge-login-ui-rn/src/components/scenes/newAccount/NewAccountTosScene.tsx
@@ -19,7 +19,7 @@ import { Checkbox } from '../../themed/Checkbox'
 import { EdgeText } from '../../themed/EdgeText'
 import { Fade } from '../../themed/Fade'
 import { SecondaryButton } from '../../themed/ThemedButtons'
-import { ThemedScene } from '../../themed/ThemedScene2'
+import { ThemedScene } from '../../themed/ThemedScene'
 
 interface OwnProps {
   branding: Branding
